requirements.txt
torch
transformers
safetensors
scikit-learn
Pillow
